requirements.txt
# LangFlix Project Dependencies

# For web scraping / transcript download
beautifulsoup4>=4.12.2
requests>=2.31.0

# For parsing .srt subtitle files
pysrt>=1.1.2

# OpenAI Python client
openai>=1.3.7

# Google Gemini API
google-generativeai>=0.8.0

# Google GenAI library for TTS
google-genai>=0.8.0

# Google Cloud Text-to-Speech (legacy, optional)
google-cloud-texttospeech>=2.16.0

# Data validation and structured output
pydantic>=2.0.0

# Configuration management
PyYAML>=6.0

# Utilities
python-dotenv>=1.0.0
inflect>=6.0.0

# Video processing
ffmpeg-python>=0.2.0
Pillow>=10.2.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
